# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GetCarApi — vehicle data collection and VIN history API
servers:
  - url: /api
    description: Base API path
...
security: []   # No global security — applied per-endpoint below
paths:
...
  /admin/vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List vehicles
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Full-text search by VIN, make, or model
        - name: make
          in: query
          schema:
            type: string
          description: Filter by make (exact, case-insensitive)
        - name: model
          in: query
          schema:
            type: string
          description: Filter by model (partial match)
        - name: yearFrom
          in: query
          schema:
            type: number
          description: Minimum model year
        - name: yearTo
          in: query
          schema:
            type: number
          description: Maximum model year
        - name: fuelType
          in: query
          schema:
            type: string
          description: Filter by fuel type
        - name: transmission
          in: query
          schema:
            type: string
          description: Filter by transmission type
        - name: providerId
          in: query
          schema:
            type: number
          description: Only vehicles seen by this provider
        - name: country
          in: query
          schema:
            type: string
          description: Filter by country (e.g. South Korea)
        - name: limit
          in: query
          schema:
            type: number
            default: 50
        - name: offset
          in: query
          schema:
            type: number
            default: 0
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleList"
  /admin/vehicles/{vin}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get a vehicle by VIN
      parameters:
        - name: vin
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vehicle with history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: VDI token (vdi_…)
      description: >
        API token issued to your account. Pass as
        `Authorization: Bearer vdi_<token>`. Required for all VIN routes
        (including check) and live inventory endpoints.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Vehicle:
      type: object
      required: [id, vin, createdAt, updatedAt]
      properties:
        id:
          type: number
        vin:
          type: string
        make:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        year:
          type: ["number", "null"]
        trim:
          type: ["string", "null"]
        bodyType:
          type: ["string", "null"]
        fuelType:
          type: ["string", "null"]
        transmission:
          type: ["string", "null"]
        driveType:
          type: ["string", "null"]
        engineDisplacement:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        currentKnownMileage:
          type: ["number", "null"]
        lastSeenAt:
          type: ["string", "null"]
          format: date-time
        listingCount:
          type: ["number", "null"]
        observationCount:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VehicleList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Vehicle"
        total:
          type: number
    VehicleObservation:
      type: object
      required: [id, vehicleId, providerId, observedAt]
      properties:
        id:
          type: number
        vehicleId:
          type: number
        providerId:
          type: number
        providerName:
          type: ["string", "null"]
        sourceListingId:
          type: ["string", "null"]
        priceAmount:
          type: ["number", "null"]
        priceCurrency:
          type: ["string", "null"]
        mileage:
          type: ["number", "null"]
        mileageUnit:
          type: ["string", "null"]
        listingStatus:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        observedAt:
          type: string
          format: date-time
    VehicleEvent:
      type: object
      required: [id, vehicleId, eventType, occurredAt]
      properties:
        id:
          type: number
        vehicleId:
          type: number
        eventType:
          type: string
        description:
          type: ["string", "null"]
        occurredAt:
          type: string
          format: date-time
    AuctionSaleRow:
      type: object
      required: [soldDate]
      properties:
        soldDate:
          type: string
          format: date
          description: Date the listing was observed or reported sold
        amount:
          type: ["number", "null"]
          description: Sale / last ask amount in minor-unscaled currency units (KRW)
        currency:
          type: ["string", "null"]
        registered:
          type: ["string", "null"]
          description: First registration date (YYYY-MM-DD)
        provider:
          type: ["string", "null"]
        sourceListingId:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
    AccidentRow:
      type: object
      required: [date, type]
      properties:
        date:
          type: string
          format: date
          description: Accident or damage date (YYYY-MM-DD)
        type:
          type: string
          enum: [accident, flood_damage, damage]
          description: Normalized accident category
        category:
          type: ["string", "null"]
          description: primary | secondary | flood | insurance type label when known
        damage:
          type: ["string", "null"]
          description: Damage area or label (e.g. Side, Front End)
        description:
          type: ["string", "null"]
        repairTotal:
          type: ["number", "null"]
          description: Total repair cost when reported (usually KRW)
        insuranceBenefit:
          type: ["number", "null"]
          description: Insurance payout when reported
        currency:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        mileageKm:
          type: ["number", "null"]
          description: Odometer in km when recorded on this accident row
        mileageMiles:
          type: ["number", "null"]
    SalvageRecord:
      type: object
      required: [salvage]
      description: >
        US/Canada title brand summary. Present only when a vehicle/auction title
        (or equivalent status) was observed. Korean market cars omit this field.
      properties:
        salvage:
          type: boolean
          description: true when title/status indicates salvage (or junk/rebuilt/COD); false for clean/clear titles
        title:
          type: ["string", "null"]
          description: Primary title text (e.g. Tx - Salvage Vehicle Title)
        detailedTitle:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
          description: Title state code when parsed (TX, CT, …)
        status:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
          format: date
        source:
          type: ["string", "null"]
    VehicleExtraRow:
      type: object
      required: [key, label, value]
      description: >
        Static auction/listing specs (keys, condition, airbags, …) — not timeline events.
        Omitted when no extra fields exist for the VIN.
      properties:
        key:
          type: string
          description: Machine-readable field id (keys, key_status, condition, …)
        label:
          type: string
          description: Human-readable label
        value:
          type: string
        source:
          type: ["string", "null"]
          description: Provider or origin (copart, iaa, salvagebid, …)
        observedAt:
          type: ["string", "null"]
          format: date
    VehicleDetail:
      allOf:
        - $ref: "#/components/schemas/Vehicle"
        - type: object
          required: [observations, events]
          properties:
            observations:
              type: array
              items:
                $ref: "#/components/schemas/VehicleObservation"
            events:
              type: array
              items:
                $ref: "#/components/schemas/VehicleEvent"
            auctionSales:
              type: array
              items:
                $ref: "#/components/schemas/AuctionSaleRow"
            accidents:
              type: array
              items:
                $ref: "#/components/schemas/AccidentRow"
            salvage:
              oneOf:
                - $ref: "#/components/schemas/SalvageRecord"
                - type: "null"
            extra:
              type: array
              items:
                $ref: "#/components/schemas/VehicleExtraRow"
              description: US auction lot specs (keys, condition, …). Omitted when empty.
            mileageHistory:
              type: array
              items:
                $ref: "#/components/schemas/MileageHistoryRow"
    MileageHistoryRow:
      type: object
      required: [date, mileageKm, mileageMiles, latest]
      description: >
        Deduplicated odometer reading collected from listings, owners, accidents,
        inspections, or any other VIN tab that recorded mileage plus a date.
        The most recent dated reading is tagged latest.
      properties:
        date:
          type: string
          format: date
        mileageKm:
          type: number
        mileageMiles:
          type: number
        kind:
          type: string
          enum: [listing, owner, accident, inspection, sale, salvage, other]
        source:
          type: ["string", "null"]
        sources:
          type: array
          items:
            type: string
        latest:
          type: boolean
        tag:
          type: string
          enum: [latest]
          description: Set to latest on the most recent dated reading

    # ── Listings ─────────────────────────────────────────────────────────────
...